Implemented a clearAllButton for favorites and implemented logic to check for an existing quote in favorites. Working on debugging the display of the heart icon when encountering a saved quote the heart icon isn't filled even though the quote is already saved

+++ js/storage.js
@@ -1,14 +1,23 @@
 // Cuvanje novog citata u localStorage
 export const saveFavoriteQuote = (quote, author) => {
     let favorites = JSON.parse(localStorage.getItem('favorites')) || [];
-    favorites.push({ quote, author });
-    localStorage.setItem('favorites', JSON.stringify(favorites));
+    // Provera da li vec postoji u omiljenim
+    const alreadyFavorite = favorites.some(fav => fav.quote === quote && fav.author === author);
+    if (!alreadyFavorite) {
+        favorites.push({ quote, author });
+        localStorage.setItem('favorites', JSON.stringify(favorites));
+    }
 };
 // Uklanjanje omiljenog citata iz localStorage
 export const removeFavoriteQuote = (quote, author) => {
     let favorites = JSON.parse(localStorage.getItem('favorites')) || [];
     favorites = favorites.filter(fav => fav.quote !== quote || fav.author !== author);
     localStorage.setItem('favorites', JSON.stringify(favorites));
+};
+// Provera da li je citat stavljen u 'omiljeno'
+export const isFavorite = (quote, author) => {
+    let favorites = JSON.parse(localStorage.getItem('favorites')) || [];
+    return favorites.some(fav => fav.quote === quote && fav.author === author);
 };
 // Prikazivanje omiljenih citata
 export const showFavoriteQuotes = () => {
@@ -19,7 +28,11 @@ export const showFavoriteQuotes = () => {
     favorites.forEach(fav => {
         const quoteDiv = document.createElement('div');
         quoteDiv.classList.add('favorite-quote');
-        quoteDiv.innerHTML = `<p>${fav.quote}</p><p>~ ${fav.author}</p>`;
+        quoteDiv.innerHTML = `<p>${fav.quote}</p><p>${fav.author}</p>`;
         favoriteQuoteList.appendChild(quoteDiv);
     })
+};
+// Funkcija za brisanje svih omiljenih citata iz localStorage
+export const clearAllFavorites = () => {
+    localStorage.removeItem('favorites');
 };
